Pass our options directly to ansible-playbook
We are not using any options that the tower setup script needs
to know about so we can avoid bugs in that script by passing
the options directly to ansible-playbook in the format it expects.

This also makes the container run the same command line as we
do on the appliance.

https://bugzilla.redhat.com/show_bug.cgi?id=1516829